Laser damage over time working

+++ Assets/Scripts/Enemy.cs
@@ -3,7 +3,7 @@
 public class Enemy : MonoBehaviour {
 
 	public float speed = 10f;
-	public int health = 100;
+	public float health = 100;
 	public int value = 50;
 
 	public GameObject deathEffect;
@@ -15,7 +15,7 @@ public class Enemy : MonoBehaviour {
 		target = Waypoints.points [0];
 	}
 
-	public void TakeDamage(int amount){
+	public void TakeDamage(float amount){
 		health -= amount;
 
 		if (health <= 0) {
